Fix simple problems in PR
