+++ jiezi/urls.py
@@ -18,6 +18,10 @@ urlpatterns = [
     # front-page urls
     path('index/', views.index, name="index"),
     path('', RedirectView.as_view(url='index')),
+    path('about_us/', views.about_us, name="about_us"),
+    path('the_science_behind/', views.the_science_behind,
+         name="the_science_behind"),
+    path('help/', views.help, name="help"),
 ]
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 urlpatterns += staticfiles_urlpatterns()
