+++ templates/base.html
@@ -71,7 +71,7 @@
 
 <div class="footer" style="background: #ccc">
         <a href="{% url 'index' %}">
-            <img src="{% static "images/base/logo.png" %}" style=" margin-left: 12vw; padding-left:3vw; height:4vw; vertical-align:middle">
+            <img src="{% static "images/logo-color-full.svg" %}" style=" margin-left: 12vw; padding-left:3vw; height:4vw; vertical-align:middle">
         </a>
         <a href="{% url 'about_us' %}" style="vertical-align: middle; margin-left: 30vw; ">
             About Us
