+++ templates/help.html
@@ -1,7 +1,6 @@
 {% extends "base.html" %}
 {% load static %}
-{% block extra_css %}
-{% endblock extra_css %}
+
 {% block title %}Help{% endblock title %}
 
 {% block content %}
