+++ templates/the_science_behind.html
@@ -1,7 +1,6 @@
 {% extends "base.html" %}
 {% load static %}
-{% block extra_css %}
-{% endblock extra_css %}
+
 {% block title %}The Science Behind{% endblock title %}
 
 {% block content %}
